refactor(ml): split open_prices.proofs.ml into submodules
- classification
- ocr
- price_tags
-receipts

This make the module easier to manage.
