+++ open_prices/proofs/management/commands/run_ml_models.py
@@ -8,11 +8,11 @@ from django.utils import timezone
 from openfoodfacts.utils import get_logger
 
 from open_prices.proofs import constants as proof_constants
-from open_prices.proofs.ml import (
+from open_prices.proofs.ml import run_and_save_proof_prediction
+from open_prices.proofs.ml.classification import PROOF_CLASSIFICATION_MODEL_NAME
+from open_prices.proofs.ml.price_tags import (
     PRICE_TAG_DETECTOR_MODEL_NAME,
-    PROOF_CLASSIFICATION_MODEL_NAME,
     run_and_save_price_tag_extraction,
-    run_and_save_proof_prediction,
 )
 from open_prices.proofs.models import PriceTagPrediction, Proof
